requirements-devel.txt
coverage
django-coverage-plugin
factory-boy
freezegun
hypothesis
pytest
pytest-cov
pytest-django
pytest-sugar
pytest-timeout
requests
